requirements.txt
Flask==3.1.1
flask-cors==4.0.0
mysql-connector-python==8.1.0